requirements.txt
# Core dependencies
agno>=2.4.0
cohere>=5.0.0
lancedb>=0.25.0

# Environment management
python-dotenv>=1.0.0

# File monitoring (for dropbox auto-ingestion)
watchdog>=3.0.0

# PDF processing
pypdf>=3.0.0

# Semantic chunking (required for SemanticChunking strategy)
chonkie[semantic]>=0.1.0
